Renamed factory header file.
